Fix navigating backwards on Special:GlobalUsage
When navigating backwards, the database result is in reversed order
compared to what is shown on the page. This was handled incorrectly
when generating the links, and as a result, instead of continuing
querying from the last row of the database result, we continued from
the last row shown on the page.

Bug: T235025
Change-Id: I0e9cb119060d111038c901c0c45ef81cf1fa31b2